Update nx_joystick.h
This update add buttons ZR, ZL and LA_STICK.  The reason is that some prefer to have block and run in Mortal Kombat with ZR and ZL and LA_STICK for "Service mode".  The update also re-order buttons so it fits properly the Mortal Kombat series mapping just like if you were on Xbox or Playstation.  It works out of the box.