New working game Ms. Pac-Man Twin
MAME WIP

0.232: Mark 'Ms Pac Man Twin (Argentina)' as a hack [hap]. Changed description to 'Ms Pac Man Twin (Argentina)'.
0.229: Roberto Fresca, Mirko Buffoni, ArcadeHacker, Rick2000, ytsejam and Recreativas.org added 'Ms PacMan Twin (Argentina)' (SUSILU 1992). 
Note: Argentine official hack of Ms. Pac-Man that allows 2 Players simultaneously. Very rare PCB. Was avoiding us for more of 22 years. 
The mainboard is a PacMan bootleg PCB, with a mini daughterboard (normal in mspacman boards) plugged in the Z80 socket.

Ali Baba and 40 Thieves

0.78u5: Pierpaolo Prazzoli added extra roms to Ali Baba. Added proms ($0, 20) and user1 rom ($0). The colors are fixed now. 
The user1 rom is unknown, the 1st and 2nd half are identical (mistery items ?).